% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_survreg_trees.r
\name{predict_survreg_trees}
\alias{predict_survreg_trees}
\title{Title}
\usage{
predict_survreg_trees(obj, newdat, x)
}
\arguments{
\item{obj}{a \code{survreg} object.}

\item{newdat}{\code{data.frame} with covariates. One of the variables must be
called \code{tdiff} and contain the time interval at which to evaluate survival
probability.}

\item{p}{vector of percentiles. See \code{predict.survreg} for details.}
}
\description{
Title
}
