% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_fpm_growth.r
\name{reg_fpm_growth}
\alias{reg_fpm_growth}
\title{Nonlinear regression for tree growth}
\usage{
reg_fpm_growth(
  df_trees,
  df_pred,
  form_expected,
  form_variance,
  species,
  country
)
}
\arguments{
\item{df_trees}{\code{data.frame} containing the sizes of every tree at two different
time points. See \code{Details} below for more information.}

\item{df_pred}{}

\item{form_expected}{formula with the expression}

\item{form_variance}{}

\item{species}{character vector containing the species for which a regression is sought.}

\item{country}{character, origin of the data set. Only \code{"spain"} has been implemented.
It can be given in lower or upper case letters, or a mixture of the two.}
}
\value{
A \code{list} whose elements correspond to each \code{species}. Each element, in turn, includes
two regression objects for the expected growth and its variance.
}
\description{
It applies a non-linear regression scheme to determine tree diameter growth as a function
of a set of predictors.
}
\details{
If the input data sets correspond to the Spanish Inventario Forestal Nacional,
the columns in \code{df_trees} must be named \code{"dbh"}, \code{"dbh1"},
\code{"species"} and \code{"tdiff"}, corresponding respectively to first and second
tree diameter, tree species and time difference (in years strictly positive) between the first and the
second time points.
}
