% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_fpm_growth.r
\name{reg_fpm_growth}
\alias{reg_fpm_growth}
\title{Nonlinear regression for tree growth}
\usage{
reg_fpm_growth(
  df_trees,
  df_pred,
  form_expected,
  form_variance,
  species,
  country
)
}
\arguments{
\item{df_trees}{}

\item{df_pred}{}

\item{form_expected}{formula with the expression}

\item{form_variance}{}

\item{species}{character vector containing the species for which a regression is sought.}

\item{country}{character, origin of the data set. Only \code{"spain"} has been implemented.
It can be given in lower or upper case letters, or a mixture of the two.}
}
\description{
It applies a non-linear regression scheme to determine tree diameter growth as a function
of a set of predictors.
}
